(iOS) Override CDVPlugin class abstract method `handleOpenURL` instead of implementing app delegate `application:openURL:options` to prevent conflicts with other plugins.
cordova-plugin-customurlscheme will not woking under current wechat plugin for ios.
That is, open the schema link on safari, the app should open a specific page. 
Now, it just open the app, rather than jump to that specific page.

I comment some code in AppDelegate+Wechat.m, and it's working again.
So I'd like to create a pull-request for this issue.